python: Move mail server and end-point config to DEFAULT_
Change strings to reference DEFAULT_ variables like the rest of the
config.

Allow the default mail servers to be set from the environment. This
isn't ideal but in the future when we set defaults based on the server
host name it will allow us to preserve using the existing arg parsing
framework to handle setting those values.