python: Option to set end-points based on server like Thunderbird
When OAUTH2_MAIL_SERVER is set this will load defaults suitable for this
host name if it is a well known host. This copies the behavior of
Mozilla Thunderbird and uses data from a snapshot of the Thunderbird
source code converted to python. The primary motivation for this fork is
to facilitate using mutt with o365. This change helps accomplish that by
providing the end-point URIs required for authentication with this type
of server while still providing backwards compatibility with
gmail-oauth2-tools.

The Thunderbird data also contains an entry for the GMail imap server so
this also continues to work with GMail even when using the new
OAUTH2_MAIL_SERVER feature.

Using an environment variable for this (OAUTH2_MAIL_SERVER) is just an
easy way to enable this feature and preserve the existing argparse
defaults.

TODO: refactor so this is the only environment variable. Move imap/smtp
server options so they are taken as arguments.